+++ restful/handler/handlers.go
@@ -253,7 +253,7 @@ type reqStruct struct {
 
 type Transaction struct {
 	Version     uint32 `json:"version"`
-	CursorNum   uint32 `json:"cursor_num"`
+	CursorNum   uint64 `json:"cursor_num"`
 	CursorLabel uint32 `json:"cursor_label"`
 	Lifetime    uint64 `json:"lifetime"`
 	Sender      string `json:"sender"`
